requirements.txt
Flask==2.0.1
requests==2.25.1
pycryptodome==3.10.1